NTPS: Requests Console Based Shipping
Ever been stuck in your office and thought "Wow, I really could go for a
burger, but the kitchen doesn't deliver." or "I'm too lazy and/or scared
to go to science to ask for a new circuitboard in person."?

Introducing NTPS: NanoTrasen Priority Shipping! All requests consoles
have been upgraded with a brand new printer, capable of producing
pre-tagged, single-use, single-object shipping packages for crew use!
Chefs can send burgers to distant diners, or science can package up that
brand new circuitboard you've been harassing them for and send it
straight to you!

But what if you have a larger parcel you want to ship, but don't have
access to cargo's coveted destination tagger? Have no fear, simply slap
the empty shipping package onto the parcel to give tag it just like
cargo!

But wait, there's more! Use a pen on any sealed shipping package and you
can include the name of your intended recipient on the package! Now,
when a coworker steals their mail, at least they will know it!

This PR adds the ability for all requests console to print off Shipping
Packages that come pre-tagged for a destination of your choosing. Each
package can hold a single item of normal or smaller size (no storage
containers, so you can't stick a bag or box in them). Once filled,
simply activate the package in hand to either dump out the item (in case
you put the wrong thing in) or seal the box for shipping. A sealed box
can then be further addressed with a pen, and dropped into any disposal
bin to start the delivery process (assuming it's connected to the
standard loop). Unsealed boxes will be delivered to disposals by way of
the cargo mail room, so make sure you seal it or hope that cargo is kind
enough to seal and resend it for you! If you have a wrapped parcel
(wrapped with packaging paper), you can also affix an empty shipping
container to the parcel to tag it yourself with the destination of the
shipping package, just in case cargo accidentally sent that virus crate
to the wrong person!