Shadowling Thrall Scaling
Shadowlings now have a scaling thrall count for acension dependent on
amount of players.
Min amount of thralls is limited to 15 as normal.
Max amount of thralls is capped to 40.
Formula is Players/2.5
Just happens 100 players is exactly 40 thralls.

Looking at a balancing standpoint, most can agree right now that
shadowlings only needing 15 thralls on 100+ rounds is way too little and
prematurely ends the round before things can get really heated up.

Also the game mode's minimum playercount is 30, to which it still
expects you to convert half the station to thralls, 15. This formula is
much less drastic than that but still raises the count.

Thanks to Kyet for writing this and explaining the in&outs of the code,
and helping me with this.