Allows shovels/spades to remove soil patches
Using a spade (or shovel from mining) on a patch of soil will do one of
two things:
- If there is a plant currently planted in the soil, it will remove it,
freeing up the soil patch.
- If there is no plant present, it will clear away the soil, deleting
the soil patch.

:cl:
rscadd: Returns the ability to remove vacant soil with shovels and
spades, and adds the ability to dig up plants in soil to remove them
instantly.
/:cl: