Rigging Lights and Cells now works with Plasma Dust
Light bulbs, light tubes, and power cells can now be rigged with plasma
dust from grinding plasma sheets as well as the standard plasma reagent
from a chem dispenser. Both reagents require at least 5u of the reagent
to be injected at once to rig the bulb/tube/cell to explode.

If a rigged cell is inserted into a stunprod, it only explodes when you
actually attempt to stun something with it, not when you turn the prod
on. This was unchanged, but I'm mentioning it so people don't think I
changed that.

Fixes #6238

:cl:
tweak: Plasma dust can now be used to rig lights and power cells, like
the plasma reagent, with explosive results.
/:cl: