Throwing Stools Works Again
Dropping a stool (not throwing) works the exact same as clicking it to
put it down (the result being a deconstructable 'structure' stool).

A thrown stool that isn't caught by a mob works the exact same as the
above (the result being a deconstructable 'structure' stool).

A thrown stool that IS caught by a mob remains in the throwable weapon
form.