Goonchat Admin Filters
This adds Message filters to Goonchat (only for mentors +
administrators).

There is a new option in the settings dropdown of Goonchat to access
these. They are temporary, non-destructive filters which will hide all
messages matching them that are already in your chat, as well as any new
messages matching them.

There are 5 filters currently:
 - Admin; Filters out most admin logs.
 - Combat; Filters out a limited subset of combat messages-
 Specifically, any message with the 'combat' span class. Currently, this
 has only been added to the central /attack and /attacked_by procs, so a
 large considerable amount of hostile actions taken against a
 player are still not going to be filtered out. We can work on adding
 the identifier to more stuff later.
 - Radios: Filters out all radio messages.
 - Speech: Filters out all mob speech.
 - OOC: Filters out OOC chat.

There is also an "All" option, which just turns off all messages that
are not internal to Goonchat.