Adds area to admin canister open notices
Modifies the "X opened a canister that contains <gas>" notice for
admins. Add the area the canister was opened in.

Reason: someone opening a toxins canister in toxins is not a big deal.
Someone opening one in central primary is. With this PR, admins can tell
the difference without having to go look at the canister.