Admins now get attack logs for punches
Right now, admins don't see attack logs for punches. They're recorded in
VV and logs, but we don't get 'ATTACK:' notices for them, even with
attack logs on.
With this PR, that changes. Now, admins with the preference enabled see
punches, too. Not disarms, just harmful punches.

We'll now be able to notice fights earlier. But, we may see a lot of
punch notices.

Admins: please use thumbs up/down on this PR according to whether or not
you want to see punches in attack logs.