SSD Handling
Adds a controller that automatically cleans up (cryos) SSD players.
After 5-10 minutes of someone being SSD, they will be automatically
moved to a free pod in cryodorms.

This process ignores anyone who is any of the following:
- Not SSD
- Dead
- Off Station
- Anchored, stunned, cuffed, buckled, or pulled
- In space
- A head of staff (including the Captain)
- Not human (ie: a simple mob, borg, etc)
- In a prison area (brig, perma)
- Has previously been moved to cryo in the same round

If you'd like to go SSD without being moved, simply buckle yourself to a
chair anywhere on the station.

In general, though, most SSD characters should get moved to cryo after a
reasonable waiting period. This helps in many ways:
- Antags with those people as objectives, promptly get new objectives.
- It frees up job slots, which can be a precious resource during
highpop.
- It reduces the number of SSD people littering the halls, thus the
amount of issues those people have, thus the amount of ahelps we have to
deal with.
- It helps avoid 'disappearing person' syndrome, where someone goes SSD
in a remote location, and you never find out what happened to them.

As a side bonus, this PR also refactors cryopod.dm to make it better.