The God Emperor lives on!
Fixes #6404

There were 2 issues here.
Firstly I'm guessing that UI has grown in length since this disk was
made... the UI on the disk only had a length of 24 instead of 32 which
was leading to a list out of bounds runtime. Fixed with a loop adding
0x000 to the end to prevent this happening again if the UI length
changes again

Secondly, even though it's a UI&UE disk, the dna record on it still
needed SE for it to work - Clean SE now added to it