Fixes Blob Conscious Split Decreasing Scaling
Makes it so that when a blob uses the Conscious Split verb, it no longer
sets it to 700 (350*2), but instead increases the current requirement by
350 (the initial amount) before the blobs can win. This gives blobs the
chance to make more cores and in turn theoretically reach that limit
quicker (and also gives them a chance to increase their play-time as
blob), but if they lose cores, they have much more work ahead of them to
make up for that loss.

Fixes #6372

:cl:Twinmold
Fix: Blob Conscious Split now properly scales the requirement of blob
tiles to win.
/:cl: