Add [OPEN] option for AI in radio messages
Also mildly refactors hear_radio to kill those nasty istype(src)'s.

This makes an [OPEN] link appear on all radio messages the AI hears (to
the right of the follow link). When clicked, it allows the AI to open
the door nearest to the speaker (or, if it is a voice changer, the door
nearest to the poor sap who had his voice stolen)