BloodPack Tagging/Pen Labeling
With this, you can now use a pen on a bloodpack and add a label to it,
giving the ability to tag a bloodpack with the type of blood from
donations. Bloodpacks will now switch between "empty" and notso in
naming, depending on if there is reagent/blood inside or not.

:cl:Twinmold
Add: Ability to label bloodpacks with a pen, so you can now label a
bloodpack from donations without a handlabeler.
Add: Bloodpacks will now change name to say if they are empty or not.
/:cl: