Tweaks Beach Awaymission
Adds a pirate area to the beach away mission.

This area contains:
- new area to explore: pirate camp
- dangerous mobs to fight: several pirates with e-cutlasses and
laserguns. Bring friends!
- loot: pirate clothing, including collectible hats, retro laser guns,
uranium, a small amount of diamond.

Note: getting to the pirate area is not easy. You have to find a hidden
diving anchor in the undersea area. The pirates themselves are
merciless. They will burst-fire lasers at you like no tomorrow. This is
NOT aimed at solo explorers!