Fixes/updates clownPlanet
Changes:
1) Clownplanet awaymission map file no longer contains a bunch of
references to objects that don't exist in the code.
2) It does, however, contain a load of bananium ore, next to the crown.

There is still no combat in this mission, and not much in the way of
items, either, except for the rare bananium. This is more a relaxed,
story mission. I'd really like to improve it, but not sure how. At least
I can make sure it doesn't contain invalid references.

🆑 Kyep
fix: Clown planet away mission is no longer horribly broken.
/🆑