Adds Employment Record Laptop to HoP's Office
Adds in an employment record laptop to the Head of Personnel's office so
they can now see people's previous employment records, making the
records much more accessible to the Head of Personnel, whom should have
easy access to these records. Removes the hand labeler andd package wrap
from the Head of Personnel's office for the space. Hand labeler is
already in the meeting room, and adds the wrap under the labeler in the
meeting room.

:cl:Twinmold
Adds: Employment record laptop to Head of Personnel's Office. Moves
package wrap that was there to meeting room to be paired with the hand
labeler there.
/:cl: