Initial version, from TG
Ports Poison Pen from TG:
https://github.com/tgstation/tgstation/pull/23778
Adds a HoP/Cargotech/QM-only traitor pen that can be used on paper to
poison it.
Picking up poisoned paper without gloves transfers the poison through
your skin, into your bloodstream. Subsequent people to pick it up are
not affected.
The poison is 15 units of the nerve agent, sarin. This is enough to kill
you in a couple of minutes. It is quite obvious.
The pen looks like a normal pen, costs 5 TC, and has 3 uses. It cannot
be used as a direct weapon. It has to be applied to paper to be
effective.

🆑 Kyep
rscadd: Added Poison Pen, a stealthy way of applying a contact-based
poison to paper. Available to Syndicate HoPs, QMs, and Cargo Techs.
/🆑