Fixes a Bug With Tail Markings
The bug occurred if a changeling assumed the identity of a character who
had a body-accessory and a tail marking specific to that body accessory
or if the person with the body accessory and marking got cloned. This
would result in a ling/player with the default tail but they'd still
have whatever marking they/the person used to have, both looking ugly
and being metagamable.