Prevents new players from attacking or looting SSDs
Changes:
- New players (those with under 10 hours of playtime) who try to attack,
or loot from, an SSD player, are prevented from doing so. Instead, they
get a message telling them to read the server rules.
- Antags are exempt from this. They can still attack/loot SSD players.
This prevents someone attacking them, then going SSD, to prevent them
retaliating.
- Since anyone with over 10 hours playtime can interact with SSDs as
normal, security officers (who typically have far more than 10 hours'
time) are typically immune also.

Why do we need this?

Because people messing with SSDs happens EVERY ROUND. And it is almost
always new players doing it. And it adds up to a big time-sink for
admins, dealing with these cases. Not to mention the hassle of cases
where it isn't caught, and someone comes back to the game to find their
ID/etc has been stolen.
We shouldn't be forcing admins to handle these cases manually, when this
code can handle 90% of the cases for us.

🆑 Kyep
rscadd: SSDs are now protected against new players messing with them.
This protection only works against new players. Anyone who has played on
the server for 10 hours or more, total, is expected to know the rules,
and is still able to interact with SSDs.
/🆑