ghetto surgery internal organ disinfection and dead limb revival
Allows to use a dropper, bottle, beaker, spray, or if you are brave
enough, an entire bucket, to treat internal organ infections with
alcohol. The more alcoholic the thing is the more it disinfects. This is
an available option during organ manipulation, at the time where you can
apply trauma kits and etc, so you can for instance apply a trauma kit
and then drip a bottle of vodka over someone's liver to treat infection.

I also noted that there is an unimplemented surgery called
fix_dead_tissue, under surgery/other.dm, which is not available in game,
so I went ahead and added it too to organ manipulation. It lets you drip
mitocholide to revive dead EXTERNAL organs, so arms and legs and chests
and etc. Does nothing to internal ones.

🆑 pinatacolada
add: ghetto surgery internal organ disinfection with alcohol
add: dead limb revival surgery step
/🆑