Sleepy Pen Price Fixed
I've changed the uplink to cost triple the telecrystals rather than just
eight for the Sleepy Pen equipment, as it is too easy for players to
cheese kills against humans, who are obviously the only species we
should be concerned about, and makes it less fun for the player getting
killed.
And you may ask, triple of 8 is 24. Well yes, this means for a traitor
to obtain such a cheese item they need a buddy willing to cough over
four more TC.