FullofSkittles Custom pAI & Silicon Fluff Tweaks
Adds in FullofSkittles's custom pAI

Adds support for custom pAI sprites

Tweaks code for custom cyborg, AI, and hologram sprites to be more
standardized, which made adding pAI support that much easier. I guess
this could also be extended later on if we wanted to support custom
maint drone sprites...

As a result of the changes, the custom_sprites.txt file will need to be
updated to the new standardized format (been working with Alffd on that
anyways so he's aware). I have taken the liberty of updating the example
custom_sprites.txt to explain and match this new format for our
downstreams and other coders.

:cl:
rscadd: Adds FullofSkittles's custom pAI sprites. Thanks FoS!
/:cl: