Removes real_name restriction for silicon sprites
Also adds code support for custom maint drones.  Whether or not we want
to let people get custom drone sprites is up to the head admins, but the
code is there for it. Maint drones have a new verb which attempts to
equip their custom sprite, if one exists for that player, which can only
be used once.