Collaborators are no longer clueless about Syndicate
Alternative to #6829

The Syndicate no longer leaves their potential collaborators in absolute
darkness about their usefulness. Collaborators are notified of their
potential use, though in a vague way so as not to compromise the
identities of any agents directly.
- Potential collaborators get messages informing them that they are
still not antags, but that they may be asked to assist, when they are
assigned to a new traitor.
- Potential collaborators get a SINGLE code word and code response so
their clandestine friend can attempt to communicate with them with more
subtlety than "hey, im a traitor, want to help?"
- They are only given 1 word set so they can't just inform security and
the AI of all the codewords alone. Same reason they aren't given a name
of the traitor they might be asked to assist.

The intercept report (currently disabled) now only lists suspects that
actually CAN be the antag (no more listing the chaplain as a potential
cultist, the HoS as a potential traitor, or that IPC in atmos as a
possible changeling).

Potential collaborators can no longer be characters in roles that are
restricted from being traitors. So no more "Potential collaborator: The
captain", as funny as that may have been.

:cl:
tweak: Potential collaborators for traitors now are informed they are
potentially a collaborator for a potential increase in them potentially
helping the traitors potentially do potentially bad things. Potentially.
rscadd: Potential collaborators are given a single code word and
response set so they can discretely find out that their best friend is a
filthy traitor.
/:cl: