Anti-Scum Safeties
All game modes EXCEPT traitor (and it's variants) misinform a number of
chumps that they are potential collaborators, to cut down on metagame
potential.

Traitor modes don't misinform anyone and instead inform the real
potential collaborators only to prevent spreading actual codewords to
the general public.

Maximum number of chumps scales with population, though it is possible
to end up with fewer chumps than the maximum based on RNG and the number
of people with opposed/skeptical relations.

Also fixed some accidentally inverted checks that broke everything, and
moved the inform_collab proc to the base game_mode type for organization