Wordlist Tweak
Removes the word 'faggot' from certain messages on the server.
Reasoning should be obvious: if using that word as a player will get you
banned, the server shouldn't be sending it to you in any automated
message.
Based on https://github.com/ParadiseSS13/Paradise/pull/6996 - but with
less controversial changes.
🆑 Kyep
tweak: Slight adjustment to word lists based on our rules.
/🆑