Nerfs the telescopic baton stun time
oh boy can you feel it, it's like I'm bathing in the ocean

Nerfs the telescopic baton stun time to be the same as it takes to cable
cuff someone, meaning you can't stuncuff them, not without help at
least, meaning it brings the telescopic baton back into the self defense
territory, meaning you can't stuncuff someone to brig, meaning it's no
longer the offensive weapon it currenty is.

It also means you can't stun lock someone, not even close to it, the
stun is like about half the cuff time.

This has been a discussion of ages, this is the most recent topic, let
the salt flow:

https://nanotrasen.se/forum/topic/9861-purge-telebatons-from-the-codebase

:cl:pinatacolada
:tweak: Nerfs the telescopic baton stun time
:cl: