Returns reagent processing checks
These control whether or not the mob should even attempt an on_mob_life
call for the reagent as well as addictions and ODs. Without this, humans
can OD on Surge and IPCs can get addicted to nicotine, which is not
intended. Whoops.

Fixes accidental re-inclusion of repurposed proc call