istype() Reduction & A Missed Spot
Checks once when S (or NS) is set. At S, it will default the species
(human), fetch the species datum (also human) and generate a random
character of that species (all but the random character generation is
done at NS as well).

Also sorts out a species check I missed (it checked to see if the
species wasn't Machine. Now it checks to see if the species doesn't have
ALL_RPARTS - functionally the same).