More changes, fixes reagent memory clearing itself
Reverts the changes to the donut New() proc

I still think this should be changed, but can't come up with a safer way
of doing it right now.

Ice cream cannot be loaded into the food replicator, preventing it from
being used to generate infinite food paste with little effort.