Oxygen Deep-crits Vox from Full-HP ~3min Quicker
Oxygen is now just as deadly as plasma to Vox (deep crit. in ~2 mins 10 seconds from full HP)

The alternative is to have max_oxy set to 0.75, which makes Vox drop into deep crit from full HP at a time just about between that which they do right now and that which they would if Oxygen was as deadly to them as plasma.

The figures:

max_oxy = time to deep crit
1	= ~05min 25sec -> Current value
0.75	= ~03min 55sec -> Middle of the road
0.5	= ~02min 25sec -> From here on, all max_oxy values kill you just about as quick.
			Might as well make it the same as plasma: 0.005.
0.05	= ~02min 15sec
0.005	= ~02min 10sec -> Just as deadly as plasma