ERTs are no longer bankrupt
ERTs now get bank accounts during their creation.
This prevents situations where the ERT is equipped with the latest
military gear, but is so penniless they can't afford a coffee from a
vending machine. More importantly, it means they won't starve if the
chef isn't making food (which, in a lot of ERT-worthy situations, they
aren't).