Admin 'Cryo SSDs' Verb
Adds a 'Cryo SSDs' verb for admins.
This takes all humans on the station who have been SSD for 10 minutes or
more, and are not immune, and puts them in the cryopods. The idea is
that it allows admins an easily way to clean up large numbers of SSD
crew on the station. SSD crew who are off-station, dead, or
pulled/restrained in some way are immune to this.

As a side benefit, this PR also refactors cryopod.dm, removing "loc ="
calls, and turning the process of entering a cryopod into a proc.