Adds 10h playtime requirement to drones
Drones are essentially mini engineering borgs, with the ability to run
under objects, ventcrawl, ignore the AI and respawn themselves. Given
that we have playtime requirements for borgs, we really should have
playtime requirements for these too.
Also, this change was requested by Neca.

🆑 Kyep
add: Playing as a drone now requires 10h of playtime.
fix: Admin-ghosts can now interact with (ie: activate) the drone fab
without having to spawn in as engineer to do so.
/ 🆑