Adds in the Jolly Jug
Due to my master exploding, I had to remake a couple of my PRs. This
pull adds in the Jolly Jug, the clown version of nothing. The clown
spawns with it and has a backup in their locker, just like the mime, in
addition it spawns with 100u of banana juice.