Atmospherics duffel bag and duffel edits
Creates a new duffel bag for Atmospherics. It is a resprite of the Engineering bag with light blue lines on it and a dark blue strap.
Also adds colored strips to each of the Medical duffels (Chem, Gene, Viro) as well as to Toxins as they seemed a little lacking in their special color. Can undo if desired as they are different from before but the same between all sprite types.

In-hand and on-mob sprites have been updated accordingly. Atmos duffel bag spawns in Atmospherics "technician's locker". There was also a missing toxins duffel bag icon for both in-hand DMIs while each other bag had one so I made those too.