Client dir compatibility, part 1
This commit modifies all of our smoothing sprites to work properly when
the client's screen is rotated. It also does the same for the lighting
overlays, adding directional sprites so that they rotate correctly.

It also ports the movement input system from Baystation, which means
that your controls aren't suddenly reversed when your screen is rotated.