Readability & Abstraction Improvements + A Tweak
As per Tigercat's review, readability has been improved and the method
by which eyeshine is applied has been abstracted a lot more.

As part of that, I have divised a solution for the situation where,
prior to this tweak, if you were a Tajara and had the 'bangs' hairstyle
while also having eyeshine, one of your eyes would render ABOVE your
hair.

Now, despite the eye shine happening on a higher layer, the eye that
would ordinarily be hidden by the hair still is.
How? By cutting the pixels that are overlapped by the hair from the eye
icon during rendering.