Temp
Fixes all the wierd ass runtimes. I think.
Dude lets be honest here, this is probably still a huge ass mess but at least now it seems to work.

I also scoured the rest of the codebase for other fixes/changes that needed to be made, including
moving Vox icon-skin-tones onto DNA proper.

Includes a bunch of optimization to species-setting and character spawning.