Grey uniform sprites (#8635)
* Adds Grey Blueshield Sprite

First ever Commit, never done this before, dunno if I've done this properly. My understanding is that all I had to do was cut the sprites into individual .pngs, import them into dream maker, assign them to the grey/species/uniform.dmi under the ert_uniform_s name (since that's the icon state name of the uniform) and since the blueshield uniform is already fully coded to use the grey uniform sprite sheet; I should only have to upload a sprite for the uniform to use.
If I've fucked this up miserably, i'm sorry.

Not a lot of guides on the actual mechanics of uploading and committing sprites/pull requests other than errorage's dense as fuck /tg/ spriting guide and guesswork.

* First Pass at Implementing the missing Grey sprites. Implements Blueshield, Pod Pilot, Psych and Coroner uniform sprites. I don't think they look great, but I realise that might be down to how strangely shaped greys are.

* Added all the Sec Loadout Uniforms and Blueshield Skirt as well as Warden Formal.

* Added Psychiatrist Turtleneck, Red and Khaki pants, Black Female Lawyer Suit, NT Rep skirt, Yellow and Blue Performers Outfits, Gladiator Outfit, Sundress and the Alternate HoP "Whimsy" uniform.
Some creative liberties have been taking with the psychiatrist's turtleneck, since it seems well-received. Given it's an alternate uniform for an underplayed job for an uncommon species, I feel okay leaving it in until someone complains and I try to butcher-code a roll up/down for it.

* Adds Sailor Dress, Presidente Suit, Soviet Uniform and Navy Blue Executive Suit. These are all the broken grey uniform sprites that i'm aware of. This includes alternative loadout uniforms, broken regular uniform sprites and uniforms you can get from the autodrobe and clothesmate.

* Fixes Yellow Performer fuck