Fixes Mech Disablers being not printable half the time
Apparently the Firedart and Mech-Disabler have the same ID for some bizarre reason. This PR fixes it. Its a minor fix and my first PR. Its basicly just changing one line to fix this pretty decent mech weapon to make it printable like it is every now and then.

Cause its weird that you can print it from round to round, but most of the time you can only print the stupid firedart.