Change windoor damage deflection to 5
makes it so most engineering tools that would logically be able to can break windoors again instead of just some of them

To give a few examples a fire extinguisher and large crowbar can break a windoor wherass a regular crowbar and a wrench can't.

after this all of the above can break windoors but using a wrench or a crowbar will take about twice as long.