Cut code duplication, added sprites
Added a new variable to magboots called slowdown_passive, letting you
set how much the magboots slow you down when not activated, by default
SHOES_SLOWDOWN (same behaviour as before). Clown magboots use
SHOES_SLOWDOWN+1, same speed as clown shoes.

Removed redefinition of toggle as the magboots can now use the original
toggle properly.

Removed var/footstep which was copy pasted from the original clown
shoes, as far as I can tell it is not used or checked anywhere and
removing it does not silence squeeks or change any other behaviour.

Added sprites for the item icon and the mob's shoes when activated and
deactivated (adds a little red light that can be seen by a keen eye when
the shoes are activated)