Window alt clicking utility
Alt clicking a window now checks both directions for availability, a clockwise rotation takes priority but when a window is already present it will rotate the window counterclockwise instead.

(prior to this it would state that the window cannot be rotated in that direction if the clockwise direction is unavailable)

https://puu.sh/Bwx9X/69079c06f2.gif A gif to show the new situation

I intentionally left out checking for the direction opposite to the current one as that would allow for an easy way to get into the station without being affected by fastmos.