SSOverlays & Hover Inventory Indicator
This ports the hover inventory indicator from /tg/, and by necessity, SSOverlays.
The hover inventory indicator basically shows where you are trying to
put an item, and then shows an item ghost in red or green, giving visual
feedback as to if that item will go in the slot or not.

SSOverlays is a much broader system for managing overlays, which should
eventually take over *all* overlays, but I have only ported the basic
system and implemented it where necessary for it to work for the hover
inventory items currently.