Adds syndicate themed internals to the syndi spacesuit box.
Buff the syndicate spacesuit by adding a syndicate mask and a dark
extended emergency oxygen tank to the syndicate spacesuit box. This
makes it easier for perma'd traitors with an uplink/storage implant to
escape, allow to carry a full EVA package in a single convenient box,
and make it easier for a spaced traitor to fix his situation.

Uplink description updated to reflect the change.