Fixes
Attempt to check for both a CLIENT and if they are SSD.

Remove from huds on new.

Var to check if they are spawned by an admin (currently just off zlevel)

Health is now 200 from 250 (any lower and i feel it may too trvial for a rare thing)

Tried to adjust prob checks on sounds, open to suggestions on values.

prob(2) on getting faster after a kill.

Attempts to check for a buckled mob and then unbuckles them on attack pahse.