Corrects some code in smite
Wasn't setting a current_victim..opps.

Other:
Attempt to prevent locker cheese, cuase of issue, its checking for a density on T(turf) first and it has no idea what to do with that.
Possible fix in code: I am checking objects on the turf, Then the turf itself first.