Adds the toxins payload casing to crafting
Adds a new type of bomb payload : The toxins payload. Crafted with a
matter bin, metal, and a remote signaller, it must be loaded with a tank
transfer valve and put into a plasteel bomb casing in order to do
anything. Once the bomb detonates, the toxins payload will open the TTV
(causing an explosion if the mix is correct).

As with syndicate bombs, a bomb with a toxin payload has a minimum 90
second timer, and will beep loudly to warn nearby crew to evacuate the
area.
It can be defused (or accidentally detonated early) in the same way a
syndicate bomb can.

This allows scientists a secure way to test their bombs (less chances of
accidental detonations with this method, although it adds extra steps),
and antagonists a way to craft their own syndicate bomb, which is better
than a simple tank transfer valve if your objective is not killing crew
but making sure a specific area is wiped off the station map with little
to no casualties.