Fancy NanoUI Improvements - Mk2 (no broke edition)
This commit steals a bunch of code from goon to make fancy nanoUI
windows behave more consistently and more pretty-ly. Hopefully, this
time around, it won't be a broken mess, like version 1 was.

The Gulpfile for Nano has been rewritten from scratch and as many old
piece of shit libraries dumped as possible. This edition comes with
actual minified files, oh my!

Shit I tested and if this breaks so help me:
 - Safe functions perfectly
 - Security records are searchable
 - syndicateMode on everything works perfectly and uses svg art by me
 (it's shit, replace when possible)
 - Cargo console works 10/10 excellently
 - Scrolling is still shit but it was more or less like it was when I
 started, so if you say "you broke it" one more fuckin time why I
 oughta!!
 - PDA and NTOS both work flawlessly (in all the Apps I could test)